# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ladino API",
    "version": "0.1.0",
    "description": "API administrativa, contable y fiscal. Errores: docs/04_PLATFORM/ERROR_CATALOG.md."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "nullable": true
          },
          "request_id": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CreatePriceListRequest": {
        "type": "object",
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "currency_code": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          }
        },
        "required": [
          "company_id",
          "name",
          "currency_code"
        ],
        "additionalProperties": false
      },
      "PriceListResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "company_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "currency_code": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "company_id",
          "name",
          "currency_code",
          "status",
          "created_at"
        ],
        "additionalProperties": false
      },
      "SetPriceRequest": {
        "type": "object",
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid"
          },
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "string",
            "pattern": "^\\d{1,16}(\\.\\d{1,8})?$"
          },
          "effective_from": {
            "type": "string",
            "format": "date-time"
          },
          "effective_to": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "company_id",
          "product_id",
          "amount",
          "effective_from"
        ],
        "additionalProperties": false
      },
      "PriceItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "price_list_id": {
            "type": "string",
            "format": "uuid"
          },
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "string",
            "pattern": "^\\d{1,16}(\\.\\d{1,8})?$"
          },
          "currency": {
            "type": "string"
          },
          "effective_from": {
            "type": "string",
            "format": "date-time"
          },
          "effective_to": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "price_list_id",
          "product_id",
          "amount",
          "currency",
          "effective_from",
          "effective_to"
        ],
        "additionalProperties": false
      }
    },
...
  },
  "paths": {
...
    "/v1/price-lists": {
      "get": {
        "summary": "Listas de precios de la empresa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Company-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Las listas.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PriceListResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sin token válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Company visible pero sin el permiso requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company o recurso no visible — indistinguible de inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Forma inválida, o company_id del cuerpo ≠ X-Company-Id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear lista de precios (permiso price_list.manage)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Company-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePriceListRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lista creada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sin token válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Company visible pero sin el permiso requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company o recurso no visible — indistinguible de inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Forma inválida, o company_id del cuerpo ≠ X-Company-Id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/price-lists/{id}/prices": {
      "get": {
        "summary": "Historial de precios; con `at` y `product_id`, el vigente A ESA FECHA",
        "description": "La fecha es parámetro, nunca el reloj del servidor (ADR-0032): un documento de ayer se recalcula con el precio de ayer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "product_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "at",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Company-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Historial (y el vigente si se pidió).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PriceItemResponse"
                      }
                    },
                    "vigente": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "amount": {
                          "type": "string"
                        },
                        "currency": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "amount",
                        "currency"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "vigente"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sin token válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Company visible pero sin el permiso requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company o recurso no visible — indistinguible de inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Forma inválida, o company_id del cuerpo ≠ X-Company-Id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cargar un precio por vigencia (append; permiso price_list.manage)",
        "description": "Un precio no se edita: corregir es una fila nueva. El período abierto anterior se cierra en el mismo INSERT (autocierre). Solape con un período cerrado → 409 PRICE_OVERLAP.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Company-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPriceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Precio cargado, {amount, currency} como strings.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceItemResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sin token válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Company visible pero sin el permiso requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company o recurso no visible — indistinguible de inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Solape con un período cerrado (PRICE_OVERLAP) o misma fecha de inicio (DUPLICATE).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Forma inválida, o company_id del cuerpo ≠ X-Company-Id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/units": {
      "get": {
        "summary": "Unidades de medida (global)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "El catálogo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "symbol": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "code",
                      "name",
                      "symbol"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sin token válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tax-categories": {
      "get": {
        "summary": "Clasificaciones tributarias activas (global, VALIDAR-TRIBUTARIO)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "El catálogo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "code",
                      "name",
                      "description",
                      "status"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sin token válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/product-categories": {
      "get": {
        "summary": "Categorías comerciales de la empresa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Company-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "El catálogo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "status"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sin token válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}